Add FrameRate and improve performance
- Run the animation at its exported frame rate
- Wrap the animation in a RepaintBoundary
- Don't paint during "static" periods
