+++ example/test/widget_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:lottie_example/main.dart';
+import 'package:lottie_example/main_app.dart';
 
 void main() {
   testWidgets('Main sample', (tester) async {
