+++ lib/src/frame_rate.dart
@@ -1,8 +1,8 @@
 class FrameRate {
   static final max = FrameRate._special(0);
-  static final fromComposition = FrameRate._special(-1);
+  static final composition = FrameRate._special(-1);
 
-  final int framesPerSecond;
+  final double framesPerSecond;
 
   FrameRate(this.framesPerSecond) : assert(framesPerSecond > 0);
   FrameRate._special(this.framesPerSecond);
