+++ lib/src/lottie_builder.dart
@@ -4,6 +4,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import '../lottie.dart';
+import 'frame_rate.dart';
 import 'lottie.dart';
 import 'providers/asset_provider.dart';
 import 'providers/file_provider.dart';
@@ -35,6 +36,7 @@ class LottieBuilder extends StatefulWidget {
     Key key,
     @required this.lottie,
     this.controller,
+    this.frameRate,
     this.animate,
     this.reverse,
     this.repeat,
@@ -46,6 +48,7 @@ class LottieBuilder extends StatefulWidget {
     this.height,
     this.fit,
     this.alignment,
+    this.addRepaintBoundary,
   })  : assert(lottie != null),
         super(key: key);
 
@@ -54,6 +57,7 @@ class LottieBuilder extends StatefulWidget {
     String src, {
     Map<String, String> headers,
     this.controller,
+    this.frameRate,
     this.animate,
     this.reverse,
     this.repeat,
@@ -67,6 +71,7 @@ class LottieBuilder extends StatefulWidget {
     this.height,
     this.fit,
     this.alignment,
+    this.addRepaintBoundary,
   })  : lottie = NetworkLottie(src,
             headers: headers, imageProviderFactory: imageProviderFactory),
         super(key: key);
@@ -84,6 +89,7 @@ class LottieBuilder extends StatefulWidget {
   LottieBuilder.file(
     Object /*io.File|html.File*/ file, {
     this.controller,
+    this.frameRate,
     this.animate,
     this.reverse,
     this.repeat,
@@ -97,6 +103,7 @@ class LottieBuilder extends StatefulWidget {
     this.height,
     this.fit,
     this.alignment,
+    this.addRepaintBoundary,
   })  : lottie = FileLottie(file, imageProviderFactory: imageProviderFactory),
         super(key: key);
 
@@ -104,6 +111,7 @@ class LottieBuilder extends StatefulWidget {
   LottieBuilder.asset(
     String name, {
     this.controller,
+    this.frameRate,
     this.animate,
     this.reverse,
     this.repeat,
@@ -119,6 +127,7 @@ class LottieBuilder extends StatefulWidget {
     this.fit,
     this.alignment,
     String package,
+    this.addRepaintBoundary,
   })  : lottie = AssetLottie(name,
             bundle: bundle,
             package: package,
@@ -129,6 +138,7 @@ class LottieBuilder extends StatefulWidget {
   LottieBuilder.memory(
     Uint8List bytes, {
     this.controller,
+    this.frameRate,
     this.animate,
     this.reverse,
     this.repeat,
@@ -142,6 +152,7 @@ class LottieBuilder extends StatefulWidget {
     this.height,
     this.fit,
     this.alignment,
+    this.addRepaintBoundary,
   })  : lottie =
             MemoryLottie(bytes, imageProviderFactory: imageProviderFactory),
         super(key: key);
@@ -159,6 +170,11 @@ class LottieBuilder extends StatefulWidget {
   /// The animated value will be mapped to the `progress` property of the
   /// Lottie animation.
   final Animation<double> controller;
+
+  /// The number of frames per second to render.
+  /// Use `FrameRate.composition` to use the original frame rate of the Lottie composition (default)
+  /// Use `FrameRate.max` to advance the animation progression at every frame.
+  final FrameRate frameRate;
 
   /// If no controller is specified, this value indicate whether or not the
   /// Lottie animation should be played automatically (default to true).
@@ -320,6 +336,13 @@ class LottieBuilder extends StatefulWidget {
   ///    relative to text direction.
   final AlignmentGeometry alignment;
 
+  /// Indicate to automatically add a `RepaintBoundary` widget around the animation.
+  /// This allows to optimize the app performance by isolating the animation in its
+  /// own `Layer`.
+  ///
+  /// This property is `true` by default.
+  final bool addRepaintBoundary;
+
   @override
   _LottieBuilderState createState() => _LottieBuilderState();
 
@@ -383,6 +406,7 @@ class _LottieBuilderState extends State<LottieBuilder> {
         Widget result = Lottie(
           composition: composition,
           controller: widget.controller,
+          frameRate: widget.frameRate,
           animate: widget.animate,
           reverse: widget.reverse,
           repeat: widget.repeat,
@@ -392,6 +416,7 @@ class _LottieBuilderState extends State<LottieBuilder> {
           height: widget.height,
           fit: widget.fit,
           alignment: widget.alignment,
+          addRepaintBoundary: widget.addRepaintBoundary,
         );
 
         if (widget.frameBuilder != null) {
