+++ lib/src/raw_lottie.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 import '../lottie.dart';
+import 'frame_rate.dart';
 import 'lottie_drawable.dart';
 import 'render_lottie.dart';
 
@@ -15,6 +16,7 @@ class RawLottie extends LeafRenderObjectWidget {
     this.delegates,
     this.options,
     double progress,
+    this.frameRate,
     this.width,
     this.height,
     this.fit,
@@ -33,6 +35,11 @@ class RawLottie extends LeafRenderObjectWidget {
 
   /// The progress of the Lottie animation (between 0.0 and 1.0).
   final double progress;
+
+  /// The number of frames per second to render.
+  /// Use `FrameRate.composition` to use the original frame rate of the Lottie composition (default)
+  /// Use `FrameRate.max` to advance the animation progression at every frame.
+  final FrameRate frameRate;
 
   /// If non-null, require the Lottie composition to have this width.
   ///
@@ -76,6 +83,7 @@ class RawLottie extends LeafRenderObjectWidget {
       delegates: delegates,
       enableMergePaths: options?.enableMergePaths,
       progress: progress,
+      frameRate: frameRate,
       width: width,
       height: height,
       fit: fit,
@@ -88,6 +96,7 @@ class RawLottie extends LeafRenderObjectWidget {
     renderObject
       ..setComposition(composition,
           progress: progress,
+          frameRate: frameRate,
           delegates: delegates,
           enableMergePaths: options?.enableMergePaths)
       ..width = width
